@@ -58,9 +58,9 @@
     "prettier": "1.9.2",
     "prettier-eslint": "^8.7.0",
     "prop-types": "^15.5.10",
-    "react": "^16.2.0",
+    "react": "^16.3.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.3.0",
     "react-hot-loader": "^3.1.3",
     "react-redux": "^5.0.6",
     "redbox-react": "^1.5.0",
@@ -88,8 +88,8 @@
     "lodash": "^4.17.4",
     "moment": "^2.18.1",
     "prop-types": "^15.5.4",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
     "redux-saga": "^0.15.6",
